fix(rules): use `uint8` instead of `int64` to store axes

+++ rules/alignment.go
@@ -10,7 +10,7 @@ import (
 
 type AlignmentRule struct {
 	Target Selector
-	Axis   int64
+	Axis   uint8
 	Radius int64 // define search radius for this rule
 }
